move BrowserNavIcon to its own directory
