+++ src/content/app/genome-browser/components/browser-nav-icon/BrowserNavIcon.scss
