+++ src/content/app/genome-browser/components/browser-nav-icon/BrowserNavIcon.test.tsx
@@ -24,7 +24,7 @@ import MockGenomeBrowser from 'tests/mocks/mockGenomeBrowser';
 
 import { BrowserNavIcon } from './BrowserNavIcon';
 
-import { browserNavConfig } from 'src/content/app/genome-browser/components/browser-nav/browserNavConfig';
+import { browserNavConfig } from 'src/content/app/genome-browser/components/browser-nav-icon/browserNavConfig';
 
 const mockGenomeBrowser = new MockGenomeBrowser();
 jest.mock(
