+++ src/content/app/genome-browser/components/browser-nav-icon/BrowserNavIcon.tsx
@@ -24,7 +24,7 @@ import ImageButton from 'src/shared/components/image-button/ImageButton';
 
 import { browserNavIconActionMap } from 'src/content/app/genome-browser/state/browser-nav/browserNavSlice';
 
-import { BrowserNavItem } from 'src/content/app/genome-browser/components/browser-nav/browserNavConfig';
+import { BrowserNavItem } from 'src/content/app/genome-browser/components/browser-nav-icon/browserNavConfig';
 import { Status } from 'src/shared/types/status';
 
 import iconStyles from './BrowserNavIcon.scss';
