+++ src/content/app/genome-browser/components/browser-nav-icon/browserNavConfig.ts
